+++ browser/UI/floatingUIPanel.js
