+++ browser/UI/panels.js
