+++ browser/UI/transmissionIncoming.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import {Entity} from 'aframe-react'
 
-const TransmissionIncoming = (props) => {
-  return (<Entity id='transmissionIncomingIndicator'
+const TransmissionIncoming = (props) => (<Entity id='transmissionIncomingIndicator'
     text={{value: 'INCOMING TRANSMISSION'}}
     position={{
       x: 0.3,
@@ -12,6 +11,5 @@ const TransmissionIncoming = (props) => {
     visible='false'
     scale={{x: 1.1, y: 1.1, z: 1.1}}
   />)
-}
 
 export default TransmissionIncoming
