+++ browser/index.js
@@ -12,8 +12,7 @@ import Navigator from './navigator'
 import Intro from './intro.jsx'
 import introText from './introText.js'
 import {mediaRecorder, startRecording, stopRecording} from './audio/audio'
-import loadAllAssets from './assets'
-import FailureView from './failureView'
+import loadAllAssets from './assets/assets'
 import { startSyncingPhaseAndStrikes } from './firebase'
 import store from './store.jsx'
 import { setNavigatorStatus, setDriverStatus } from './reducers/strike-phase.js'
