+++ browser/tutorial/introText.js
